test: relocated plugin manifest tests in correct namespace
